Updated View.js to handle canvas' in iframes
Remember each canvas' ownerDocument. If the canvas' ownerDocument is the document then add DomEvent.add events to the iframe. This is needed when the iframe obscures some or all of the document. On remove view verify that there are no other views referencing the same iframe before removing the event handlers.