Loosen text for existing canvas
Presume a canvas is good even if it doesn't pass the "instanceof HTMLCanvasElement" test AS LONG as the document and the canvas.ownerDocument are different. Presume the caller knows what they're doing when iframes are involved. Maybe a bad assumption - there might be a better way to test - we could get a 2d context and try to write to it, but that seems like a lot of trouble.