Fix item drawing within a clipped group
When item was part of a clipped group and a child of a group with matrix
not applied that had transformations and when item needed to be drawn in
a separated canvas (partial opacity, special blend mode, ...), item
drawing was incorrectly clipped.
Drawing clip item before applying transformations to the context in
`Item#draw()` fixes this problem.
Closes #1594